setting up dev server, starting auth routes
