+++ package.json
@@ -5,17 +5,18 @@
   "type": "module",
   "main": "index.js",
   "scripts": {
-    "dev": "ts-node src/index.ts"
+    "dev": "tsx watch src/index.ts"
   },
   "author": "",
   "license": "ISC",
-  "dependencies": {
-    "@prisma/client": "^5.13.0",
-    "express": "^4.19.2"
-  },
+  "dependencies": {},
   "devDependencies": {
+    "tsx": "^4.7.3",
+    "@prisma/client": "^5.13.0",
     "@types/express": "^4.17.21",
     "@types/node": "^20.12.7",
+    "express": "^4.19.2",
+    "nodemon": "^3.1.0",
     "prisma": "^5.13.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5"
