+++ src/routes/user.routes.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import { registerUser } from "../controllers/user.controller.js";
+
+const router = Router();
+
+router.route("/register").post(registerUser);
+
+router.get("/login", () => {});
+
+export default router;
